Improved move sort

+++ algorithm/move_sort.h
@@ -5,6 +5,6 @@
 
 void clear_killer_moves();
 void add_killer_move(uint16_t move, uint8_t depth);
-uint16_t get_mvvlva(BoardState *from, BoardState *to);
+uint8_t get_mvvlva(BoardState *from, BoardState *to);
 void sort_moves(BoardState *from, BoardStack *stack, uint16_t base, uint16_t tt_move, uint8_t depth);
 void sort_moves_q(BoardState *from, BoardStack *stack, uint16_t base);
